refactor: condense master RM modules together.
Move several pieces of code into `resourcemanagers` because they belong to it:
- `internal/agent`
- `internal/kubernetes`
- `internal/provisioner`

Also a couple small code quality improvements:
- Move StartTaskPod/KillTaskPod from sproto into kubernetes because
  they're effectively private kubernetes RM messages, not used
  elsewhere.
- Remove `resourcemanagers.container#agent` field as it's effectively unused.
