+++ master/internal/resourcemanagers/agent/slot.go
