+++ master/internal/resourcemanagers/agent/slots.go
