+++ master/internal/resourcemanagers/provisioner/agent_setup_test.go
@@ -11,7 +11,7 @@ import (
 )
 
 func TestAgentSetupScript(t *testing.T) {
-	err := etc.SetRootPath("../../static/srv/")
+	err := etc.SetRootPath("../../../static/srv/")
 	assert.NilError(t, err)
 
 	encodedScript := base64.StdEncoding.EncodeToString([]byte("sleep 5\n echo \"hello world\""))
