+++ master/internal/resourcemanagers/provisioner/aws_spot.go
