+++ master/internal/resourcemanagers/provisioner/aws_spot_collections.go
