+++ master/internal/resourcemanagers/provisioner/aws_spot_legacy.go
